se mejora la estructura para el /
