+++ desafios/implementacion de login/.env
@@ -1,2 +1,3 @@
-MONGO_URL=mongodb+srv://coderTutor:3g7zdMaTQoelzFS3@cluster0.koltvse.mongodb.net/ecommerce
+MONGO_URL=mongodb+srv://:@cluster0.xxxxx.mongodb.net/ecommerce
 PORT=8080
+SECRET_PHRASE=secret
