+++ desafios/implementacion de login/package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/dacerb/full-stack-coderhouse-c43395/tree/primeraEntrega",
   "dependencies": {
+    "connect-mongo": "^5.0.0",
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.3.1",
     "eslint": "^8.48.0",
