+++ desafios/implementacion de login/src/routes/front.apiRoutes.js
@@ -1,0 +1,9 @@
+import * as APIController from "../controllers/apiControllers.js";
+import router from "./front.cartRoutes.js";
+
+
+// router.get("/", ProductController.getProducts); old
+router.get("/", APIController.getHome);
+
+
+export default router;
